Support for weblate version 4.7
This change is needed to fetch changes from weblate 4.7

'changes' must end with '/'as '.../api/changes/'  or it will
return 404 (Page Not Found).

There are 'unit' URLs that return 404 (Not Found) when they were
removed. It is now a WARNING instead of an ERROR to keep running.

Test updated accordingly.

Signed-off-by: Quan Zhou <quan@bitergia.com>